fix: 주문 확정 시 고객 통계 자동 업데이트 시스템 구현

See merge request a701-be/a701-be!55

+++ src/main/java/com/kkulddip/order/application/service/PriceValidationException.java
@@ -15,7 +15,7 @@ public class PriceValidationException extends BusinessException {
     /**
      * 가격 불일치 예외
      */
-    public static PriceValidationException priceMismatch(Long productId, Integer expectedPrice, Integer actualPrice) {
+    public static PriceValidationException priceMismatch(Long productId, Long expectedPrice, Long actualPrice) {
         String message = String.format("상품 가격이 일치하지 않습니다. 상품ID: %d, 요청 가격: %d, 실제 가격: %d", 
             productId, expectedPrice, actualPrice);
         return new PriceValidationException(message);
@@ -32,7 +32,7 @@ public class PriceValidationException extends BusinessException {
     /**
      * 잘못된 단가 예외
      */
-    public static PriceValidationException invalidUnitPrice(Long productId, Integer unitPrice) {
+    public static PriceValidationException invalidUnitPrice(Long productId, Long unitPrice) {
         String message = String.format("잘못된 단가입니다. 상품ID: %d, 단가: %d", productId, unitPrice);
         return new PriceValidationException(message);
     }
